refactor: merge component, render, oob, pjx_load into reactive/reactive.py
ReactiveComponent, the render descriptor, reactive_render_bundle, oob_swaps,
and the PjxLoad field marker now live in one cohesive module. Behavior unchanged.
